+++ devenv.nix
@@ -5,5 +5,5 @@
   inputs,
   ...
 }: {
-  packages = [pkgs.git pkgs.rustup pkgs.openssl pkgs.cargo-nextest pkgs.taplo pkgs.protobuf];
+  packages = [pkgs.git pkgs.rustup pkgs.openssl pkgs.cargo-nextest pkgs.taplo pkgs.cargo-insta pkgs.protobuf];
 }